feat: represent Square x and y values using nibbles
Replace the x and y fields in the Square struct with a single val field
that holds both values in the same u8. Use bit shifting and bitwise
operations to extract the x and y values from the val field. This
change reduces the size of the Square struct and simplifies some of the
code.

BREAKING CHANGE: The Square struct's x and y fields have been replaced
with a single val field. Code that uses the old fields will need to be
updated. All code that used the old fields in square.rs have already
been updated.